requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
# PostgreSQL - Render uchun (Linux da o'rnatiladi)
psycopg2-binary==2.9.9
